Add Elaborate phase, annotations
This adds an elaborate phase. This requires the addition of two new
annotations:
  - ChiselGeneratorAnnotation storing a () => RawModule
  - ChiselCircuitAnnotation storing a Chisel Circuit

The ChiselGeneratorAnnotation object includes an additional apply
method to allow this to be safely constructed from a String, e.g., if
constructed via a command line argument. Safe construction means that
expected, possible exceptions are wrapped in an OptionsException with
a sane message.

The actual elaboration is done in an "elaborate" method of
ChiselGeneratorAnnotation. The elaboration phase then just calls this
method for any discovered ChiselGeneratorAnnotations to create
ChiselCircuitAnnotations while deleting the original.
OptionsExceptions are rethrown while ChiselExceptions use the Driver's
stack trace pruning logic and just delete the ChiselGeneratorAnnotation.

This adds tests of the individual annotations and of the Elaborate
phase.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>